requirements.txt
requests
python-dotenv
duckdb
pandas
matplotlib
seaborn
scikit-learn
boto3
ijson
s3fs
fsspec
logging
wordcloud
numpy
prefect